Changes JsonArrayBindingSource to utilize generic
This changes the JsonArrayBindingSource to utilize a generic type
parameter. This should make it easier to work with the class and
reduce the need for casting.
A non-generic class was also created that inherits from the generic
class with JsonValue in the parameter. This is for backwards
compatability.
Some small amount of code cleanup is also included in these changes.